Changed seatingPlanId to long due to unhandled exception if integer is overflowed

+++ backend/src/main/java/at/ac/tuwien/sepm/groupphase/backend/endpoint/dto/ShowSearchDto.java
@@ -34,7 +34,7 @@ public class ShowSearchDto   {
   private BigDecimal price;
 
   @JsonProperty("seatingPlan")
-  private Integer seatingPlan;
+  private Long seatingPlan;
 
   public ShowSearchDto event(String event) {
     this.event = event;
@@ -93,7 +93,7 @@ public class ShowSearchDto   {
     this.price = price;
   }
 
-  public ShowSearchDto seatingPlan(Integer seatingPlan) {
+  public ShowSearchDto seatingPlan(Long seatingPlan) {
     this.seatingPlan = seatingPlan;
     return this;
   }
@@ -104,11 +104,11 @@ public class ShowSearchDto   {
   */
   
   @Schema(name = "seatingPlan", required = false)
-  public Integer getSeatingPlan() {
+  public Long getSeatingPlan() {
     return seatingPlan;
   }
 
-  public void setSeatingPlan(Integer seatingPlan) {
+  public void setSeatingPlan(Long seatingPlan) {
     this.seatingPlan = seatingPlan;
   }
